Make cache result computation synchronized.
Multiple threads may want the same freshly initialized cache entry. It is guaranteed to be unique by the previous sync on the cache, however, the result itself is not covered by the cache's lock. No lock means computing the same twice and dropping one of the results, as reference assignments are atomic, but the more obvious, the better.